backend/requirements.txt
Django
Djangorestframework
djangorestframework-simplejwt
django-cors-headers
